Lab 4
This lab required the use of sub-queries to form the queries to the questions. The operations that I mostly used in this lab were the intersect operator, order by, and select distinct. The first 8 questions were relatively simple, I struggled slightly with questions 9-10. I was able to produce the answer that corresponds correctly to the cap database through due time.