Lab 10 Submission
Didn't struggle to badly with this one, when I tried to make the input courseNum int, it wouldn't take it as it said was too vague. So I changed it to course_Num int and it worked. I had to look up how to return a table of info, so I made both functions return the coursenum and coursename that the respective function relates to.